new design for MCPToolAction
