+++ openhands/sdk/mcp/__init__.py
@@ -1,7 +1,7 @@
 """MCP (Model Context Protocol) integration for agent-sdk."""
 
 from openhands.sdk.mcp.client import MCPClient
-from openhands.sdk.mcp.definition import MCPToolObservation
+from openhands.sdk.mcp.definition import MCPToolAction, MCPToolObservation
 from openhands.sdk.mcp.tool import (
     MCPTool,
     MCPToolExecutor,
@@ -14,6 +14,7 @@ from openhands.sdk.mcp.utils import (
 __all__ = [
     "MCPClient",
     "MCPTool",
+    "MCPToolAction",
     "MCPToolObservation",
     "MCPToolExecutor",
     "create_mcp_tools",
