+++ openhands/sdk/tool/tool.py
@@ -195,7 +195,7 @@ class Tool(DiscriminatedUnionMixin, Generic[ActionT, ObservationT]):
         self,
         add_security_risk_prediction: bool = False,
     ) -> ChatCompletionToolParam:
-        """Convert an MCP tool to an OpenAI tool.
+        """Convert Tool to an OpenAI tool for LLM completion.
 
         Args:
             add_security_risk_prediction: Whether to add a `security_risk` field
